pyproject.toml
[tool.poetry]
name = "newsman"
version = "0.1.0"
description = "A news API built with FastAPI"
authors = ["Daniel Boadzie <boadziedaniel43@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
tortoise-orm = "^0.17.4"
email-validator = "^1.1.3"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.910"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
